[issue #4] proper timestamp update for npm/bower install
grunt.task.run was misused as it does not execute tasks right away but
enqueues them. So the assumption that one can directly check the success
of the task execution was simply wrong. Now, an additional task is
introduced that is only executed if the previous tasks did not fail.